+++ packages/ag-charts-community/src/module-bundles/all.ts
@@ -1,6 +1,6 @@
 import { type ModuleDefinition } from 'ag-charts-core';
 
-import { GaugePresetModule, PriceVolumePresetModule, SparklinePresetModule } from '../api/preset/presetModules';
+import { SparklinePresetModule } from '../api/preset/presetModules';
 import { AllCartesianCommunityModules } from './cartesian';
 import { AllPolarCommunityModules } from './polar';
 
@@ -9,7 +9,5 @@ export const AllCommunityModule: ModuleDefinition[] = [
     ...AllPolarCommunityModules,
 
     // Presets
-    PriceVolumePresetModule,
-    GaugePresetModule,
     SparklinePresetModule,
 ];
