+++ packages/ag-charts-enterprise/src/module-bundles/all.ts
@@ -1,10 +1,11 @@
-import { AllCommunityModule, GaugePresetModule, PriceVolumePresetModule } from 'ag-charts-community';
+import { AllCommunityModule } from 'ag-charts-community';
 import type { ModuleDefinition } from 'ag-charts-core';
 
 import { AllCartesianEnterpriseModules } from './cartesian';
 import { AllPolarEnterpriseModules } from './polar';
 import { AllStandaloneEnterpriseModules } from './standalone';
 import { AllTopologyEnterpriseModules } from './topology';
+import { GaugePresetModule, PriceVolumePresetModule } from '../preset/presetModules';
 
 export const AllEnterpriseModule: ModuleDefinition[] = [
     ...AllCommunityModule,
