+++ packages/ag-charts-enterprise/src/preset/gauge.ts
