+++ packages/ag-charts-enterprise/src/preset/priceVolumePreset.ts
@@ -20,9 +20,11 @@ import type {
     DatumDefault,
 } from 'ag-charts-types';
 
-import type { ChartTheme } from '../../chart/themes/chartTheme';
-import { SAFE_STROKE_FILL_OPERATION } from '../../chart/themes/util';
+import { _ModuleSupport, _Theme } from 'ag-charts-community';
 import { annotationsTheme } from './priceVolumePresetTheme';
+
+type ChartTheme = _Theme.ChartTheme;
+const { SAFE_STROKE_FILL_OPERATION } = _ModuleSupport;
 
 const chartTypes = ['ohlc', 'line', 'step-line', 'hlc', 'high-low', 'candlestick', 'hollow-candlestick'];
