+++ packages/ag-charts-enterprise/src/preset/priceVolumePresetTheme.ts
@@ -6,23 +6,27 @@ import type {
     WithThemeParams,
 } from 'ag-charts-types';
 
-import { FONT_SIZE_RATIO } from '../../chart/themes/constants';
-import {
-    DEFAULT_ANNOTATION_HANDLE_FILL,
-    DEFAULT_ANNOTATION_STATISTICS_COLOR,
-    DEFAULT_ANNOTATION_STATISTICS_DIVIDER_STROKE,
-    DEFAULT_ANNOTATION_STATISTICS_DOWN_FILL,
-    DEFAULT_ANNOTATION_STATISTICS_DOWN_STROKE,
-    DEFAULT_ANNOTATION_STATISTICS_FILL,
-    DEFAULT_ANNOTATION_STATISTICS_STROKE,
-    DEFAULT_FIBONACCI_STROKES,
-    DEFAULT_FINANCIAL_CHARTS_ANNOTATION_BACKGROUND_FILL,
-    DEFAULT_FINANCIAL_CHARTS_ANNOTATION_COLOR,
-    DEFAULT_TEXTBOX_COLOR,
-    DEFAULT_TEXTBOX_FILL,
-    DEFAULT_TEXTBOX_STROKE,
-    DEFAULT_TEXT_ANNOTATION_COLOR,
-} from '../../chart/themes/symbols';
+import { _ModuleSupport } from 'ag-charts-community';
+
+const {
+    FONT_SIZE_RATIO,
+    ThemeSymbols: {
+        DEFAULT_ANNOTATION_HANDLE_FILL,
+        DEFAULT_ANNOTATION_STATISTICS_COLOR,
+        DEFAULT_ANNOTATION_STATISTICS_DIVIDER_STROKE,
+        DEFAULT_ANNOTATION_STATISTICS_DOWN_FILL,
+        DEFAULT_ANNOTATION_STATISTICS_DOWN_STROKE,
+        DEFAULT_ANNOTATION_STATISTICS_FILL,
+        DEFAULT_ANNOTATION_STATISTICS_STROKE,
+        DEFAULT_FIBONACCI_STROKES,
+        DEFAULT_FINANCIAL_CHARTS_ANNOTATION_BACKGROUND_FILL,
+        DEFAULT_FINANCIAL_CHARTS_ANNOTATION_COLOR,
+        DEFAULT_TEXTBOX_COLOR,
+        DEFAULT_TEXTBOX_FILL,
+        DEFAULT_TEXTBOX_STROKE,
+        DEFAULT_TEXT_ANNOTATION_COLOR,
+    },
+} = _ModuleSupport;
 
 const stroke = {
     stroke: DEFAULT_FINANCIAL_CHARTS_ANNOTATION_COLOR,
